# speedup: Sort alternates, to extract more common prefixes