--- a/go.mod
+++ b/go.mod
@@ -1,3 +1,3 @@
 module github.com/grafana/regexp
 
-go 1.17
+go 1.21