--- a/syntax/parse.go
+++ b/syntax/parse.go
@@ -5,6 +5,7 @@
 package syntax
 
 import (
+	"slices"
 	"sort"
 	"strings"
 	"unicode"
@@ -591,6 +592,28 @@ func (p *parser) collapse(subs []*Regexp, op Op) *Regexp {
 func (p *parser) factor(sub []*Regexp) []*Regexp {
 	if len(sub) < 2 {
 		return sub
+	}
+
+	// Sort literals, to bring similar strings nearer to each other.
+	if len(sub) > 2 {
+		allHaveLeadingString := true
+		for i := range sub {
+			if x, _ := p.leadingString(sub[i]); x == nil {
+				allHaveLeadingString = false
+				break
+			}
+		}
+		if allHaveLeadingString {
+			// Where one string is a prefix of the other we want to keep the ordering the same,
+			// because earlier choices are preferred. So use stable sort, and only compare up
+			// to the length of the shorter string.
+			slices.SortStableFunc(sub, func(a, b *Regexp) int {
+				lsa, _ := p.leadingString(a)
+				lsb, _ := p.leadingString(b)
+				pref := min(len(lsa), len(lsb))
+				return slices.Compare(lsa[:pref], lsb[:pref])
+			})
+		}
 	}
 
 	// Round 1: Factor out common literal prefixes.