The context for this optimisation is where a long list of alternates is supplied, similar to the case in the benchmark.
Such lists can be generated by Grafana dashboard variables, although that particular case is already optimized
in Prometheus' `FastRegexMatcher`.

We still see regexps with long lists of alternatives coming from other automated processes, that slip past those optimizations.  Sorting inside the regexp package allows `factor()` to extract more common prefixes, which makes matching much more efficient.

Example: `abc|x|abd` - sorting allows the `ab` prefix to be extracted.
One test needs to be amended because of this.

Where one alternate is already a prefix of another, avoid moving the shorter one earlier in the list (where it would be preferred), by using stable sort and comparing up to the shorter length.

Need to bump Go language version to 1.21, so I can use `min()` and the `slices` package.

Benchmark:
```
goos: darwin
goarch: arm64
pkg: github.com/grafana/regexp
             │  before.txt  │             after.txt              │
             │    sec/op    │   sec/op     vs base               │
Alternates-8   11.253µ ± 2%   1.100µ ± 0%  -90.22% (p=0.002 n=6)
```