@bboreham 2024-05-27
Upstream: https://go-review.googlesource.com/c/go/+/588535