@buro9 2024-06-03
conjecture that reverse sorting may be preferred as it preserves specificity in the matching whilst also resulting in a sorted tree